More texture blending

+++ src/slap42/graphics/meshes/entity_mesh.cpp
@@ -5,7 +5,7 @@
 
 namespace Slap42 {
 
-EntityMesh::EntityMesh(float* vertices, int vertices_size_bytes, unsigned short* indices, int indices_size_bytes)
+EntityMesh::EntityMesh(float* vertices, size_t vertices_size_bytes, unsigned short* indices, size_t indices_size_bytes)
     : Mesh(vertices, vertices_size_bytes, indices, indices_size_bytes) {
   GL_CHECK(glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 5 * sizeof(float), (void*)(0 * sizeof(float))));
   GL_CHECK(glEnableVertexAttribArray(0));
